@@ -2,7 +2,7 @@
   "author": "Ilan Laloum",
   "license": "MIT",
   "name": "ng-wiki-project",
-  "version": "1.0.0",
+  "version": "1.0.6",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/ilanlal/ng-wiki.git"
@@ -18,34 +18,34 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "~13.2.0",
-    "@angular/cdk": "^13.2.6",
-    "@angular/common": "~13.2.0",
-    "@angular/compiler": "~13.2.0",
-    "@angular/core": "~13.2.0",
-    "@angular/forms": "~13.2.0",
-    "@angular/material": "^13.2.6",
-    "@angular/platform-browser": "~13.2.0",
-    "@angular/platform-browser-dynamic": "~13.2.0",
-    "@angular/router": "~13.2.0",
+    "@angular/animations": "~13.3.0",
+    "@angular/cdk": "^13.3.0",
+    "@angular/common": "~13.3.0",
+    "@angular/compiler": "~13.3.0",
+    "@angular/core": "~13.3.0",
+    "@angular/forms": "~13.3.0",
+    "@angular/material": "^13.3.0",
+    "@angular/platform-browser": "~13.3.0",
+    "@angular/platform-browser-dynamic": "~13.3.0",
+    "@angular/router": "~13.3.0",
     "ng-packagr": "^13.3.0",
-    "ng-wiki": "^1.0.4",
-    "rxjs": "~7.5.0",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.11.4"
+    "ng-wiki": "^1.0.5",
+    "rxjs": "~7.5.5",
+    "tslib": "^2.3.1",
+    "zone.js": "~0.11.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.6",
-    "@angular/cli": "~13.2.4",
-    "@angular/compiler-cli": "~13.2.0",
-    "@types/jasmine": "~3.10.0",
-    "@types/node": "^12.11.1",
-    "jasmine-core": "~4.0.0",
-    "karma": "~6.3.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.1.0",
-    "karma-jasmine": "~4.0.0",
+    "@angular-devkit/build-angular": "~13.3.0",
+    "@angular/cli": "~13.3.0",
+    "@angular/compiler-cli": "~13.3.0",
+    "@types/jasmine": "~4.0.0",
+    "@types/node": "^17.0.22",
+    "jasmine-core": "~4.0.1",
+    "karma": "~6.3.17",
+    "karma-chrome-launcher": "~3.1.1",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.5.2"
+    "typescript": "~4.6.2"
   }
 }
